refactor(pipeline-core): relocate core steps as well
