+++ src/cli/repl/server/connection.ts
@@ -26,9 +26,9 @@ import { deepMergeObject } from '../../../util/objects'
 import { LogLevel } from '../../../util/log'
 import { printStepResult, StepOutputFormat } from '../../../core/print/print'
 import { DataflowInformation } from '../../../dataflow/internal/info'
-import { PARSE_WITH_R_SHELL_STEP } from '../../../core/steps/all/00-parse'
-import { NORMALIZE } from '../../../core/steps/all/10-normalize'
-import { LEGACY_STATIC_DATAFLOW } from '../../../core/steps/all/20-dataflow'
+import { PARSE_WITH_R_SHELL_STEP } from '../../../core/steps/all/core/00-parse'
+import { NORMALIZE } from '../../../core/steps/all/core/10-normalize'
+import { LEGACY_STATIC_DATAFLOW } from '../../../core/steps/all/core/20-dataflow'
 
 /**
  * Each connection handles a single client, answering to its requests.
