+++ test/functionality/pipelines/create/create-tests.ts
@@ -1,10 +1,10 @@
 import { createPipeline } from '../../../../src/core/steps/pipeline'
 import { IStep, NameOfStep } from '../../../../src/core/steps'
 import { expect } from 'chai'
-import { PARSE_WITH_R_SHELL_STEP } from '../../../../src/core/steps/all/00-parse'
+import { PARSE_WITH_R_SHELL_STEP } from '../../../../src/core/steps/all/core/00-parse'
 import { allPermutations } from '../../../../src/util/arrays'
-import { NORMALIZE } from '../../../../src/core/steps/all/10-normalize'
-import { LEGACY_STATIC_DATAFLOW } from '../../../../src/core/steps/all/20-dataflow'
+import { NORMALIZE } from '../../../../src/core/steps/all/core/10-normalize'
+import { LEGACY_STATIC_DATAFLOW } from '../../../../src/core/steps/all/core/20-dataflow'
 import { STATIC_SLICE } from '../../../../src/core/steps/all/static-slicing/30-slice'
 import { NAIVE_RECONSTRUCT } from '../../../../src/core/steps/all/static-slicing/40-reconstruct'
